+++ components/Gallery/index.tsx
@@ -18,6 +18,7 @@ import Background from '../Background';
 import ClothApi from '@/apis/domain/Cloth/ClothApi';
 import { ColorListType } from '../ColorList';
 import { suggestionColorType } from '@/pages/add-cloth';
+import { findClosestColor, kMeansClustering } from '@/utils/ColorExtraction';
 
 interface GalleryProps {
   imageAndTag: ImageWithTag | undefined;
@@ -217,115 +218,6 @@ const Gallery = ({
     }
   };
 
-  function kMeansClustering(pixels: RGB[], k: number): Cluster[] {
-    let centroids = initializeCentroids(pixels, k);
-    let clusters: Cluster[] = [];
-
-    for (let i = 0; i < 10; i++) {
-      clusters = assignPixelsToClusters(pixels, centroids);
-      centroids = updateCentroids(clusters);
-    }
-
-    return clusters;
-  }
-
-  function initializeCentroids(pixels: RGB[], k: number): RGB[] {
-    const centroids: RGB[] = [];
-    for (let i = 0; i < k; i++) {
-      const randomPixel = pixels[Math.floor(Math.random() * pixels.length)];
-      centroids.push(randomPixel);
-    }
-    return centroids;
-  }
-
-  function assignPixelsToClusters(pixels: RGB[], centroids: RGB[]): Cluster[] {
-    const clusters: Cluster[] = centroids.map((centroid) => ({
-      centroid,
-      pixels: [],
-    }));
-
-    for (let pixel of pixels) {
-      let minDistance = Number.MAX_VALUE;
-      let closestCentroidIndex = -1;
-
-      for (let i = 0; i < centroids.length; i++) {
-        const distance = colorDistance(pixel, centroids[i]);
-        if (distance < minDistance) {
-          minDistance = distance;
-          closestCentroidIndex = i;
-        }
-      }
-
-      clusters[closestCentroidIndex].pixels.push(pixel);
-    }
-
-    return clusters;
-  }
-
-  function updateCentroids(clusters: Cluster[]): RGB[] {
-    return clusters.map((cluster) => {
-      const sum = cluster.pixels.reduce(
-        (acc: RGB, pixel: RGB) => {
-          acc.r += pixel.r;
-          acc.g += pixel.g;
-          acc.b += pixel.b;
-          return acc;
-        },
-        { r: 0, g: 0, b: 0 }
-      );
-
-      const count = cluster.pixels.length;
-      return { r: sum.r / count, g: sum.g / count, b: sum.b / count };
-    });
-  }
-
-  function hexToRgb(hex: string): RGB {
-    let bigint = parseInt(hex.slice(1), 16);
-    return {
-      r: (bigint >> 16) & 255,
-      g: (bigint >> 8) & 255,
-      b: bigint & 255,
-    };
-  }
-
-  function colorDistance(color1: RGB, color2: RGB): number {
-    return Math.sqrt(
-      Math.pow(color1.r - color2.r, 2) +
-        Math.pow(color1.g - color2.g, 2) +
-        Math.pow(color1.b - color2.b, 2)
-    );
-  }
-
-  function findClosestColor(
-    r: number,
-    g: number,
-    b: number,
-    colorList: ColorListType
-  ): suggestionColorType | null {
-    if (!colorList || colorList.length === 0) return null; // default to null if color list is empty
-
-    let minDistance = Number.MAX_VALUE;
-    let closestColorName: suggestionColorType | null = null;
-
-    const targetColor = { r, g, b };
-
-    for (let color of colorList) {
-      let colorRgb = hexToRgb(color.colorCode);
-      let distance = colorDistance(targetColor, colorRgb);
-
-      if (distance < minDistance) {
-        minDistance = distance;
-        closestColorName = {
-          id: color.id,
-          name: color.name,
-          colorCode: color.colorCode,
-        }; // Return the name of the closest color
-      }
-    }
-
-    return closestColorName;
-  }
-
   return (
     <>
       <Background
@@ -365,7 +257,6 @@ const Gallery = ({
             <Body4 className="selected" state="emphasis">
               {selectedImage.length}장의 사진이 선택됨
             </Body4>
-            <h1>color : {suggestionColor && suggestionColor.colorCode}</h1>
             <Carousel
               infinite={false}
               dots={true}
